feat: Products views
The /products/new view renders the form to create a new product.
The form, contains a field_for "products[]", by this way, is possible to
insert multiple products in the same request, then the controller catch
this array and permit all the parameters.

Also was created the resources for products.